// webapp/src/shared/generated/models/UpdateChatroomDto.ts
/* tslint:disable */
/* eslint-disable */
/**
 * transcendence-app
 * The transcendence-app API description
 *
 * The version of the OpenAPI document: 1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { exists, mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface UpdateChatroomDto
 */
export interface UpdateChatroomDto {
    /**
     * 
     * @type {string}
     * @memberof UpdateChatroomDto
     */
    name?: string;
    /**
     * 
     * @type {string}
     * @memberof UpdateChatroomDto
     */
    oldPassword?: string | null;
    /**
     * 
     * @type {string}
     * @memberof UpdateChatroomDto
     */
    password?: string | null;
    /**
     * 
     * @type {string}
     * @memberof UpdateChatroomDto
     */
    confirmationPassword?: string | null;
    /**
     * 
     * @type {number}
     * @memberof UpdateChatroomDto
     */
    avatarX?: number;
    /**
     * 
     * @type {number}
     * @memberof UpdateChatroomDto
     */
    avatarY?: number;
}

/**
 * Check if a given object implements the UpdateChatroomDto interface.
 */
export function instanceOfUpdateChatroomDto(value: object): boolean {
    let isInstance = true;

    return isInstance;
}

export function UpdateChatroomDtoFromJSON(json: any): UpdateChatroomDto {
    return UpdateChatroomDtoFromJSONTyped(json, false);
}

export function UpdateChatroomDtoFromJSONTyped(json: any, ignoreDiscriminator: boolean): UpdateChatroomDto {
    if ((json === undefined) || (json === null)) {
        return json;
    }
    return {
        
        'name': !exists(json, 'name') ? undefined : json['name'],
        'oldPassword': !exists(json, 'oldPassword') ? undefined : json['oldPassword'],
        'password': !exists(json, 'password') ? undefined : json['password'],
        'confirmationPassword': !exists(json, 'confirmationPassword') ? undefined : json['confirmationPassword'],
        'avatarX': !exists(json, 'avatarX') ? undefined : json['avatarX'],
        'avatarY': !exists(json, 'avatarY') ? undefined : json['avatarY'],
    };
}

export function UpdateChatroomDtoToJSON(value?: UpdateChatroomDto | null): any {
    if (value === undefined) {
        return undefined;
    }
    if (value === null) {
        return null;
    }
    return {
        
        'name': value.name,
        'oldPassword': value.oldPassword,
        'password': value.password,
        'confirmationPassword': value.confirmationPassword,
        'avatarX': value.avatarX,
        'avatarY': value.avatarY,
    };
}